pyproject.toml
[project]
name = "genai-aws-circleci"
version = "0.1.0"
description = "Building a Serverless GenAI API with FastAPI, AWS, and CircleCI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
]